add products price test case
