+++ react-shop-test/src/app/TestWrapper.ts
@@ -1,0 +1,19 @@
+import {QueryClient, QueryClientProvider} from "react-query";
+import {Wrapper} from "react-test-wrapper/react-testing-library";
+
+const queryClientOption = {
+  defaultOptions: {
+    queries: {
+      retry: false,
+      cache: false
+    },
+  },
+};
+const queryClient = new QueryClient(queryClientOption);
+
+const WrapperComponent = new Wrapper(QueryClientProvider)
+  .withDefaultProps({
+    client: queryClient
+  })
+
+export { WrapperComponent }
